+++ src/common.ts
@@ -1,0 +1,3 @@
+export const getRandomNumber = (range: number) => {
+  return Math.floor(Math.random() * range);
+};
