+++ src/util.ts
@@ -1,0 +1,5 @@
+import { getRandomNumber } from './common';
+
+export const getRandomItem = (inventory: string[]) => {
+  return inventory[getRandomNumber(inventory.length)];
+};
